platform/agents/requirements.txt
hagworm==5.9.16
requests==2.31.0
aiohttp==3.8.4
aiofiles==23.2.1
html5tagger==1.3.0
redis==3.5.3
redis-py-cluster==2.1.3
sanic==23.6.0
websocket-client==1.6.4
